+++ src/Domain/GetJobsResponse.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace DjThossi\ErgosoftSdk\Domain;
+
+readonly class GetJobsResponse
+{
+    public function __construct(
+        public StatusCode $statusCode,
+        public JobCollection $jobs,
+        public GetJobsResponseBody $responseBody,
+    ) {
+    }
+}