+++ src/Domain/GetJobsResponseBody.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace DjThossi\ErgosoftSdk\Domain;
+
+readonly class GetJobsResponseBody extends JsonResponseBody
+{
+}